# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dopamine Market API
  description: >
    Синтетический маркетплейс/сервис доставки — портфолио-проект.
    Монетизация: фиксированная плата 10₽ за оформленный заказ (INV-01).
    Полностью синтетический каталог и курьеры, симуляция доставки —
    таймер-стейт-машина (см. docs/prd/PRD.md, ADR-005, ADR-009).
  version: "1.0.0"
  contact:
    name: Dopamine Market (портфолио-проект)
servers:
  - url: http://localhost:8080
    description: Локальная разработка
  - url: https://api.dopamine-market.example.com
    description: Продакшн (заглушка домена)
...
paths:
...
  /orders:
    post:
      tags: [order]
      summary: Оформление заказа из текущей корзины
      description: Фиксированная стоимость 10.00 RUB (INV-01). Очищает корзину и генерирует событие order.created.v1 (ADR-003).
      operationId: createOrder
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "201":
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Конфликт — уже есть активный заказ в процессе оплаты (INV-02)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      tags: [order]
      summary: Список заказов текущего пользователя
      operationId: listOrders
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /orders/stats:
    get:
      tags: [order]
      summary: Статистика пользователя и streak заказов (EPIC-14, FR-GAMIFY-01)
      operationId: getUserStats
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: tz
          in: query
          description: Имя часового пояса клиента (IANA time zone, например Europe/Moscow)
          schema:
            type: string
            example: "Europe/Moscow"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_orders:
                    type: integer
                  current_streak_days:
                    type: integer
                  longest_streak_days:
                    type: integer
                  total_saved_rub:
                    type: string
                  last_order_at:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /orders/{id}/cancel:
    post:
      tags: [order]
      summary: Отмена заказа пользователем
      operationId: cancelOrder
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    PickupPoint:
      type: object
      required: [id, name, latitude, longitude, distance_meters, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        distance_meters:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
...
    OrderItem:
      type: object
      required: [product_id, name, category_name, price_rub, image_seed, quantity, subtotal_rub]
      properties:
        product_id:
          type: string
          format: uuid
        name:
          type: string
        category_name:
          type: string
        price_rub:
          type: string
          example: "10.00"
        image_seed:
          type: string
        quantity:
          type: integer
        subtotal_rub:
          type: string
          example: "10.00"
    Order:
      type: object
      required: [id, user_id, status, total_amount_rub, pickup_point, items, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, payment_pending, paid, assembling, courier_assigned, in_transit, delivered, payment_failed, cancelled]
          example: "created"
        total_amount_rub:
          type: string
          example: "10.00"
        pickup_point:
          $ref: "#/components/schemas/PickupPoint"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...